[update to 4.2] read update notes or readme file
I've been without motivation for this project for quite some time now. I'm considering dropping everything and either starting anew or otherwise rebuilding it all from the ground up. The previous branch, "main" will still be hosted at the same address with the commit previous to the one this update is pushed to for the time being. HoYo have their own calculator of sorts already. Other websites will have similar tools to this.